quantitative_analysis_project/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
alpha_vantage>=2.3.1
statsmodels>=0.14.0
scipy>=1.10.0
pyyaml>=6.0
torch>=2.0.0
scikit-learn>=1.3.0
ta>=0.10.0
fredapi>=0.5.0
pytest>=7.0.0